docs: add CONTRIBUTING.md file
Add CONTRIBUTING.md file to help to get started the development
environment, the day to day tasks when making changes on the repository
and guidelines about the commit message format.

Some information at CONTRIBUTING.md already was in README.md; that
existing information is removed from README.md

We hope this helps on keeping consistency (even for us) and make the
repository maintenance easier for maintainer and contributors.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>