refactor: adapt to mockery v3
Make the necessary changes to adapt to mockery v3:

- Migrate the configuration in a .mockery.yml as showed
  in the documentation, allowing more control, and decouple
  from the makefile rules.
- Update Makefile rule to use the configuration file.
- Update tools to use mockery v3
- Update .gitignore to add .mockery.yml file to the repo.

Mockery allow to define the mock generation in a more
modular way by using this file.

If the filename has _test.go suffix, it cannot be imported
in other places where the mock is required.

We modify the filename globally, so the code generated for
the mocks can be imported from the place where the unit
tests are implemented, to isolated from external dependencies.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>